backend/pyproject.toml
[project]
name = "unmsm-admission-analytics-backend"
version = "0.1.0"
description = "Backend foundation for UNMSM Admission Analytics"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "fastapi>=0.135.1,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "PyJWT>=2.10.1,<3.0.0",
  "boto3>=1.40.55,<2.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "redis>=5.2.1,<6.0.0",
  "uvicorn[standard]>=0.41.0,<1.0.0",
  "pydantic-settings>=2.13.1,<3.0.0",
  "sqlalchemy>=2.0.48,<3.0.0",
  "alembic>=1.18.4,<2.0.0"
]

[build-system]
requires = ["setuptools>=82.0.1", "wheel>=0.46.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
